Lab1: fixed segfaults
Encrypt and decrypt no longer segfault. Decrypt seems to not be working
properly, maybe an issue with how it checks for the end of the file and start of
the MAC. It needs to handle a blank file in which case it will not read the
number of bytes that it is expecting to.